% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-reg-glue.R
\name{add_registration_glue}
\alias{add_registration_glue}
\title{This examines a package swift file and builds the necessary registration glue code}
\usage{
add_registration_glue(package = ".")
}
\arguments{
\item{package}{where the package lives}
}
\value{
nothing directly, but has a side effect of creating \code{init.c}
}
\description{
This examines a package swift file and builds the necessary registration glue code
}
