% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiftr-package.R
\docType{package}
\name{swiftr}
\alias{swiftr}
\alias{swiftr-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/swiftr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/swiftr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
